* Add HyphenatorFactory * Maintain hyphenator settings in Hyphenator only, not directly in   voikko_options.
git-svn-id: https://voikko.svn.sourceforge.net/svnroot/voikko/trunk@3070 ee56993d-3414-0410-8ed4-8136e77a0dd8

+++ libvoikko/src/hyphenator/AnalyzerToFinnishHyphenatorAdapter.hpp
@@ -43,7 +43,7 @@ class AnalyzerToFinnishHyphenatorAdapter : public Hyphenator {
 		bool uglyHyphenation;
 		bool hyphenateUnknown;
 		int intersectCompoundLevel;
-		int minHyphenatedWordLength;
+		size_t minHyphenatedWordLength;
 		bool ignoreDot;
 		
 		/**
